move generateVerificationCode util from messages to utils
